removed the unecessary try catchs as Carlos mentioned from TopHeader
also implemented a common method to create an opportunit, in this case it is a static methods, I don't think
that it si neede to instante a class for this.

in EditOpportunity test you can see the usage. In this case it is using the JSON taht I created,for the other cases you can create the simple ones as we were discussion yesterday

please let me know what you guys think.